requirements.txt
Flask==0.10.1
Flask-Admin==1.0.7
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
SQLAlchemy==0.9.4
Werkzeug==0.9.4
alembic==0.6.4
oslo.config==1.3.0
pbr==0.8.0
